+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/client_secret.py
@@ -4,8 +4,7 @@
 # ------------------------------------
 from typing import TYPE_CHECKING
 
-from .base import AsyncCredentialBase
-from .._internal import AadClient
+from .._internal import AadClient, AsyncContextManager
 from .._internal.decorators import log_get_token_async
 from ..._internal import ClientSecretCredentialBase
 
@@ -14,7 +13,7 @@ if TYPE_CHECKING:
     from azure.core.credentials import AccessToken
 
 
-class ClientSecretCredential(AsyncCredentialBase, ClientSecretCredentialBase):
+class ClientSecretCredential(AsyncContextManager, ClientSecretCredentialBase):
     """Authenticates as a service principal using a client ID and client secret.
 
     :param str tenant_id: ID of the service principal's tenant. Also called its 'directory' ID.
@@ -24,6 +23,10 @@ class ClientSecretCredential(AsyncCredentialBase, ClientSecretCredentialBase):
     :keyword str authority: Authority of an Azure Active Directory endpoint, for example 'login.microsoftonline.com',
           the authority for Azure Public Cloud (which is the default). :class:`~azure.identity.AzureAuthorityHosts`
           defines authorities for other clouds.
+    :keyword bool enable_persistent_cache: if True, the credential will store tokens in a persistent cache. Defaults to
+          False.
+    :keyword bool allow_unencrypted_cache: if True, the credential will fall back to a plaintext cache when encryption
+          is unavailable. Default to False. Has no effect when `enable_persistent_cache` is False.
     """
 
     async def __aenter__(self):