+++ sdk/storage/azure-storage-blob/tests/test_blob_client.py
@@ -382,6 +382,7 @@ class StorageClientTest(StorageTestCase):
             self.assertTrue(service.primary_endpoint.startswith('https://www.mydomain.com/'))
             self.assertTrue(service.secondary_endpoint.startswith('https://www-sec.mydomain.com/'))
 
+
     def test_create_service_with_custom_account_endpoint_path(self):
         account_name = "blobstorage"
         account_key = "blobkey"
@@ -437,6 +438,31 @@ class StorageClientTest(StorageTestCase):
         self.assertEqual(service.credential, None)
         self.assertEqual(service.primary_hostname, 'local-machine:11002/custom/account/path')
         self.assertEqual(service.url, 'http://local-machine:11002/custom/account/path/foo/bar?snapshot=baz')
+
+    def test_create_blob_client_with_sub_directory_path_in_blob_name(self):
+        blob_url = "https://testaccount.blob.core.windows.net/containername/dir1/sub000/2010_Unit150_Ivan097_img0003.jpg"
+        blob_client = BlobClient.from_blob_url(blob_url)
+        self.assertEqual(blob_client.container_name, "containername")
+        self.assertEqual(blob_client.blob_name, "dir1/sub000/2010_Unit150_Ivan097_img0003.jpg")
+
+        blob_emulator_url = 'http://127.0.0.1:1000/devstoreaccount1/containername/dir1/sub000/2010_Unit150_Ivan097_img0003.jpg'
+        blob_client = BlobClient.from_blob_url(blob_emulator_url)
+        self.assertEqual(blob_client.container_name, "containername")
+        self.assertEqual(blob_client.blob_name, "dir1/sub000/2010_Unit150_Ivan097_img0003.jpg")
+
+    def test_create_client_for_emulator(self):
+        container_client = ContainerClient(
+            account_url='http://127.0.0.1:1000/devstoreaccount1',
+            container_name='newcontainer',
+            credential='Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==')
+
+        self.assertEqual(container_client.container_name, "newcontainer")
+        self.assertEqual(container_client.account_name, "devstoreaccount1")
+
+        ContainerClient.from_container_url('http://127.0.0.1:1000/devstoreaccount1/newcontainer')
+        self.assertEqual(container_client.container_name, "newcontainer")
+        self.assertEqual(container_client.account_name, "devstoreaccount1")
+
 
     @GlobalStorageAccountPreparer()
     def test_request_callback_signed_header(self, resource_group, location, storage_account, storage_account_key):