+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,11 +1,20 @@
 # Release History
 
-## 3.0.0b1 (Unreleased)
+## 3.0.0b2 (Unreleased)
+
+**New features**
+
+- Client-level, keyword argument `api_version` can be used to specify the service API version to use. Currently only v2.0
+is supported. See the enum `FormRecognizerApiVersion` for supported API versions.
+- `FormWord` and `FormLine` now have attribute `kind` which specifies the kind of element it is, e.g. "word" or "line"
+
+## 3.0.0b1 (2020-08-11)
 
 The version of this package now targets the service's v2.0 API.
 
 **Breaking Changes**
 
+- Client library version bumped to `3.0.0b1`
 - Values are now capitalized for enums `FormContentType`, `LengthUnit`, `TrainingStatus`, and `CustomFormModelStatus`
 - `document_name` renamed to `name` on `TrainingDocumentInfo`
 - Keyword argument `include_sub_folders` renamed to `include_subfolders` on `begin_training` methods